sql: create local trait to convert Value+WithOptionValue to other types
